+++ app/src/main/java/com/ytowka/worktimer2/utils/timers/coroutine/CoroutineActionTimer.kt
@@ -3,4 +3,4 @@ package com.ytowka.worktimer2.utils.timers.coroutine
 import com.ytowka.worktimer2.data.models.Action
 import com.ytowka.worktimer2.utils.timers.ActionTimer
 
-class CoroutineActionTimer(override val action: Action) : CoroutineTimer(action.duration.toLong(), action.exactTimeDefine), ActionTimer
+class CoroutineActionTimer(override val action: Action) : CoroutineTimer(action.duration, action.exactTimeDefine), ActionTimer
